python_service/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
selenium==4.18.1
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
webdriver-manager==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.1 